requirements.txt
line-bot-sdk
flask
beautifulsoup4==4.9.3
requests==2.25.0